Article for the "Encontros Universitários" of 2021
My article, for the "Encontros Universitários" of the Federal University of Ceará, with the subject of Computer Architecture, which concerns my scholarship i.e. monitoring, where I talk about how, and why, the teaching of assembly language is a key concept in the construction of critical thinking in computer programming.